Added more code to UDP layer
Mostly just added code to remove most of the warnings being generated
when the code was compiled, to make relevant warning messages more
identifiable. All of the code added should ultimately be usable once the
transition to the new stack is complete, it was just simple code changes
like writing getters and setters for header fields and adding a small
tcp file so that TCPPacket can be included in the TransportPacket enum.